requirements.txt
asgiref==3.9.1
django==4.2.23
django-filter==25.1
djangorestframework==3.16.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing-extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
